p1: adapt var to store 1e10

+++ 1/1.cpp
@@ -5,7 +5,7 @@ int main()
     int cases;
     std::cin >> cases;
     
-    int a, b;
+    long a, b;
     for(int i = 0; i < cases; i++)
     {
         std::cin >> a >> b;
